remove jsonl commands from CI config
