+++ .circleci/config.yml
@@ -290,9 +290,6 @@ jobs:
     executor: blockframes-ci
     steps:
       - prepare_environment
-      - run: npm run db:download-prod
-      - run: npm run db:anonymize
-      - run: npm run db:upload-anon-db-ci
       - run: npm run firestore:run-db-anon-process
   copy-storage-from-prod:
     executor: blockframes-ci
