+++ package.json
@@ -51,9 +51,6 @@
     "bundlesize": "bundlesize",
     "postinstall": "ngcc --properties es2015 browser module main --first-only --create-ivy-entry-points && node ./decorate-angular-cli.js",
     "prerender:festival": "nx build festival --prod && nx run festival:server --configuration production && angular-prerender --browser-target festival:build --server-target festival:server",
-    "db:download-prod": "ts-node -P tools/tsconfig.tools.json tools/scripts/get-prod-backup.ts",
-    "db:anonymize": "ts-node -P tools/tsconfig.tools.json tools/scripts/anonymize-db.ts",
-    "db:upload-anon-db-ci": "ts-node -P tools/tsconfig.tools.json tools/scripts/upload-anon-db-ci.ts",
     "firestore:run-db-anon-process": "npm run backend-ops anonProdDb",
     "backend-ops": "ts-node -P tools/tsconfig.tools.json apps/backend-ops/src/main.ts"
   },
