--- tools/scripts/restore-storage-from-ci.ts
@@ -1,5 +1,0 @@
-import 'tsconfig-paths/register';
-import { loadAdminServices, restoreStorageFromCi } from '@blockframes/firebase-utils';
-
-const adminServices = loadAdminServices();
-restoreStorageFromCi(adminServices.getCI());
